pages added to routes

+++ src/main/java/com/codecool/restflights/Controller/PassengersRestController.java
@@ -4,6 +4,7 @@ import com.codecool.restflights.Model.Passenger;
 import com.codecool.restflights.Service.Intarfaces.PassengerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
@@ -24,11 +25,20 @@ public class PassengersRestController {
     @DeleteMapping("/{id}")
     @Transactional
     public void deletePassenger(@PathVariable(name = "id") long id){
+        Passenger passenger = passengerService.findPassengerByPassengerId(id);
+        if (passenger == null){
+            throw new ResourceNotFoundException();
+        }
+
         passengerService.deletePassengerByPassengerId(id);
     }
 
     @GetMapping("/{id}")
     public Passenger getPassenger(@PathVariable(name = "id") long id){
+        Passenger passenger = passengerService.findPassengerByPassengerId(id);
+        if (passenger == null){
+            throw new ResourceNotFoundException();
+        }
 
         return passengerService.findPassengerByPassengerId(id);
     }
